requirements.txt
Django==1.11.7
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.12.2
automium==0.2.6
click==6.7
djangorestframework==3.7.3
itsdangerous==0.24
legacy==0.1.6
mysqlclient==1.3.9
pip==9.0.1
pytz==2017.3
setuptools==36.7.0
wheel==0.30.0
IPy==0.83.0